[python] Share strict Arrow import validation
